The test harness now displays nice tracebacks when tests fail, if it can.
I originally tried using the exception service for this, with code like this:

    let excSvc = Components.classes["@mozilla.org/exceptionservice;1"].
                 getService(Components.interfaces.nsIExceptionService);
    let excMgr = excSvc.currentExceptionManager;
    let exception = excMgr.getCurrentException();

but the exception was always null, so that didn't work.

One big caveat here is that the traceback info will only be displayed if the exception caught is either a nsIException (in which case the 'location' nsIStackFrame property is examined) or a bona fide JS exception (in which case the 'stack' property is examined).  If, however, the exception is merely a string, it can't get any traceback info.  So this is a good reason for promoting the throwing of actual exceptions instead of bare strings, e.g.:

  throw new Error("Something bad happened!")

instead of

  throw "Something bad happened!"

The latter is the way most weave exceptions are thrown, so we should probably change it.